# Make version.Compare more efficient