This also fixes the issue with intransitive compares.

Net change:

```
benchmark              old ns/op     new ns/op     delta
BenchmarkCompare-8     10311121      426626        -95.86%

benchmark              old allocs     new allocs     delta
BenchmarkCompare-8     176620         836            -99.53%

benchmark              old bytes     new bytes     delta
BenchmarkCompare-8     18729708      40641         -99.78%
```